Refactor GameClient and Tools for improved command handling and introduce start_next_combat method
- Simplified asynchronous command execution in GameClient by streamlining task handling.
- Added start_next_combat method in Tools for initiating combat, enhancing game command functionality.
- Updated GameRoot to improve port handling by using int parsing for better validation.

+++ Scripts/GameRoot.cs
@@ -25,7 +25,7 @@ public partial class GameRoot : Node
 	{
 		commandHandler = new(this);
 		if (arguments.TryGetValue(key: "port", value: out var portText))
-			if (ushort.TryParse(s: portText, result: out var port))
+			if (int.TryParse(s: portText, result: out var port))
 			{
 				Log.Print($"启动服务器，端口: {port}");
 				mcpHandler = new(gameRoot: this, port: port);
